refactor(ui): break up TransactionsScreen into smaller focused files
TransactionsScreen.kt was a massive 2,867-line file that was difficult to
navigate and maintain. This refactoring splits it into 6 focused files in
a new transactions subpackage:

- TransactionComponents.kt (~150 lines) - Shared utilities and helpers
- TransactionCard.kt (~220 lines) - Individual transaction card component
- TransactionEntryDialog.kt (~470 lines) - New transaction creation form
- TransactionEditDialog.kt (~580 lines) - Transaction editing form
- TransactionAuditScreen.kt (~750 lines) - Audit history viewer
- TransactionsScreen.kt (~900 lines) - Main screen with matrix and list

Benefits:
- Improved maintainability with single-responsibility files
- Easier navigation and code discovery
- Reduced merge conflict risk
- Maximum file size reduced from 2,867 to ~900 lines

Closes #270
